comments changes and tests + database changes (waiting confirmation to finish it)

+++ controllers/comments.controller.js
@@ -1,6 +1,7 @@
 const db = require('../models/index');
 const Comment = db.comment
 const messages = require('../utilities/messages');
+const validation = require('../utilities/validation.js');
 
 exports.create = async (req, res) => {
     try {
@@ -12,6 +13,8 @@ exports.create = async (req, res) => {
         } else {
             comment.activityId = req.params.idA
         }
+
+        comment.userId = req.loggedUser.id
         
 
         let newComment = await Comment.create(comment)
@@ -27,18 +30,21 @@ exports.create = async (req, res) => {
 
 exports.edit = async (req, res) => {
     try {
-
         let comment = await Comment.findByPk(req.params.commentId)
-
-        if(req.body.description != comment.description) {
-            comment.description = req.body.description
-            comment.edited = true
+        console.log(req.loggedUser.id)
+        if(req.loggedUser.id == comment.userId){
             
-            res.status(200).json({
-                sucess: true,
-                msg: `Comment updated successfully`
-              })
+            if(req.body.description != comment.description) {
+                comment.description = req.body.description
+                comment.edited = true
+                
+                res.status(200).json({
+                    sucess: true,
+                    msg: `Comment updated successfully`
+                  })
+            }
         }
+
 
     } catch (err) {
         res.status(500).json(messages.errorInternalServerError());
@@ -53,7 +59,7 @@ exports.findAll = async (req, res) => {
             comments = await Comment.findAll({where: {eventId: req.params.id}})
             res.status(200).json(comments)
         } else {
-            comments = await Comment.findAll({where: {eventId: req.params.idA}})
+            comments = await Comment.findAll({where: {activityId: req.params.idA}})
             res.status(200).json(comments)
         }
 
@@ -64,10 +70,9 @@ exports.findAll = async (req, res) => {
 
 exports.findOne = async (req, res) => {
     try {
-        let comment
+        let comment = await Comment.findByPk(req.params.commentId)
         
-        if(req.params.id != null ){
-            comment = await Comment.findAll(req.params.commentId)
+        // if(req.params.id != null ){
 
             if(comment != undefined){
                 res.status(200).json(comment)
@@ -75,15 +80,14 @@ exports.findOne = async (req, res) => {
                 res.status(400).json({succes: false, message: 'Invalid comment.'})
             }
 
-        } else {
-            comment = await Comment.findAll(req.params.commentId)
+        // } else {
 
-            if(comment != undefined){
-                res.status(200).json(comment)
-            } else {
-                res.status(400).json({succes: false, message: 'Invalid comment.'})
-            }
-        } 
+            // if(comment != undefined){
+            //     res.status(200).json(comment)
+            // } else {
+            //     res.status(400).json({succes: false, message: 'Invalid comment.'})
+            // }
+        // } 
 
     } catch (err) {
         res.status(500).json(messages.errorInternalServerError());
@@ -92,11 +96,15 @@ exports.findOne = async (req, res) => {
 
 exports.delete = async (req, res) => {
     try {
-        let comment = await Comment.destroy({where: {id: req.params.commentId}})
-        res.status(200).json({
-            sucess: true,
-            msg: `Comment deleted successfully`,
-          })
+        let comment = await Comment.findByPk(req.params.commentId)
+        
+        if(req.loggedUser.id == comment.userId) {
+            await Comment.destroy({where: {id: req.params.commentId}})
+            res.status(200).json({
+                sucess: true,
+                msg: `Comment deleted successfully`,
+              })
+        }
     } catch (err) {
         res.status(500).json(messages.errorInternalServerError());
     }
@@ -105,11 +113,13 @@ exports.delete = async (req, res) => {
 exports.rating = async (req, res) => {
     try {
         let comment = await Comment.findByPk(req.params.commentId)
-
+        console.log(comment.like)
         if(req.body.rating == "like") {
-            comment.like += 1;
+                comment.like = comment.like + 1;
+                res.status(200).send({msg: `Comment liked`, rating: comment.like})
         } else if (req.body.rating == "dislike"){
             comment.dislike += 1;
+            res.status(200).send({msg: `Comment disliked`})
         }
 
     } catch (err) {
